meeple/meeple/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-nested-admin==4.1.1
django-nested-formset==0.1.4
mysqlclient==2.2.7
psycopg==3.2.4
psycopg2==2.9.10
python-dotenv==1.0.1
python-monkey-business==1.1.0
sqlparse==0.5.3
